+++ test/flutter_store_test.dart
@@ -3,7 +3,17 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:atomic_state/atomic_state.dart';
 import 'package:flutter_atomic_state/flutter_atomic_state.dart';
 
-import 'fixtures/counter.dart';
+class Counter {
+  final int count;
+
+  Counter({this.count = 0});
+  factory Counter.from(Counter counter, {int count}) =>
+      Counter(count: count ?? counter.count);
+
+  Counter inc([int value = 1]) => Counter.from(this, count: count + value);
+
+  Map<String, dynamic> toJson() => {'count': count};
+}
 
 void main() {
   testWidgets('flutter widgets', (WidgetTester tester) async {
@@ -11,19 +21,19 @@ void main() {
 
     await tester.pumpWidget(StoreProvider(
         store: store,
-        child: StoreConnector<Store>(
-          builder: (context, store) => Text(
-              store.getState<Counter>().count.toString(),
-              key: Key('count'),
-              textDirection: TextDirection.ltr),
-        )));
+        child: StoreConnector<Counter>(
+            builder: (context, store, counter) => GestureDetector(
+                onTap: () =>
+                    store.updateState<Counter>((counter) => counter.inc()),
+                child: Text('${counter.count}',
+                    key: Key('count'), textDirection: TextDirection.ltr)),
+            converter: (store) => store.getState())));
 
-    store.updateState<Counter>((counter) => counter.inc());
-
+    await tester.tap(find.byType(GestureDetector));
     await tester.pump();
 
     final text = tester.firstWidget<Text>(find.byKey(Key('count')));
 
-    expect(int.parse(text.data), greaterThan(0));
+    expect(int.parse(text.data), equals(1));
   });
 }