+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## [0.1.1] - Default Converter bug fix
+## [0.1.3] - API changes
+
+Pass store to builder, require converter
+
+## [0.1.2] - Default Converter bug fix
 
 Default Converter function bug fix for invalid type.